update accountant to use u128 for calculations

+++ programs/accountant/src/state/generic_accountant.rs
@@ -38,13 +38,13 @@ impl Accountant for GenericAccountant {
     }
 
     fn enter(&self, amount: u64) -> Result<u64> {
-        let fee = self.entry_fee * amount / FEE_BPS;
-        Ok(fee)
+        let fee = self.entry_fee as u128 * amount as u128 / FEE_BPS as u128;
+        Ok(fee as u64)
     }
 
     fn redeem(&self, amount: u64) -> Result<u64> {
-        let fee = self.redemption_fee * amount / FEE_BPS;
-        Ok(fee)
+        let fee = self.redemption_fee as u128 * amount as u128 / FEE_BPS as u128;
+        Ok(fee as u64)
     }
 
     fn distribute(&mut self, accounts: &Distribute) -> Result<()> {
